PYIC-1718: Rename the session-start lambda
