+++ lambdas/initialise-ipv-session/src/main/resources/IpvLambdaJsonLayout.json
