+++ lambdas/initialise-ipv-session/src/main/resources/log4j2.yaml
