+++ settings.gradle
@@ -7,7 +7,7 @@ include "lib",
 		"lambdas:get-credential-issuer-config",
 		"lambdas:evaluate-gpg45-scores",
 		"lambdas:build-debug-credential-data",
-		"lambdas:sessionstart",
+		"lambdas:initialise-ipv-session",
 		"lambdas:useridentity",
 		"lambdas:process-journey-step",
 		"lambdas:build-cri-oauth-request",
